# doc/api-documentation.yml
openapi: 3.0.3
info:
  title: Chocolate Customizer API
  description: API for chocolate customization service and product management
  version: 1.1.0
servers:
  - url: http://localhost:3000/api/v1
    description: Production API path (Default port 3000 if not specified)
  - url: http://localhost:3030/api/v1
    description: Production API path (Default port 3030 if not specified in env)
components:
  schemas:
...
    Admin:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: "admin"
        firstName:
          type: string
          example: "Admin"
        lastName:
          type: string
          example: "User"
        email:
          type: string
          format: email
          example: "admin@example.com"
        role:
          type: string
          example: "admin"
    LoginCredentials:
      type: object
      properties:
        username:
          type: string
          example: "admin"
        password:
          type: string
          format: password
          example: "password"
    UnifiedLoginCredentials:
      type: object
      properties:
        email:
          type: string
          description: Email for customer or username for admin
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "password"
    PasswordResetRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: Email address for password reset
          example: "user@example.com"
    PasswordReset:
      type: object
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
          description: Password reset token sent to user's email
          example: "password-reset-token-123"
        newPassword:
          type: string
          format: password
          description: New password to set
          example: "newSecurePassword123!"
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        token:
          type: string
          example: "mock-jwt-token"
        user:
          type: object
          properties:
            id:
              type: integer
            username:
              type: string
            role:
              type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Error message"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
paths:
...
  /admin/login:
    post:
      summary: Admin login
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginCredentials'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/profile:
    get:
      summary: Get admin profile
      operationId: getAdminProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Admin'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update admin profile
      operationId: updateAdminProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Admin'
                  message:
                    type: string
                    example: "Profile updated successfully"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/unified-login:
    post:
      summary: Unified login for customers and admins
      operationId: unifiedLogin
      description: Authenticates users with email/password, automatically detecting whether they are customers or admin users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnifiedLoginCredentials'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/validate:
    post:
      summary: Validate user account
      operationId: validate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  description: Validation token sent to user's email
                  example: "abc123def456"
      responses:
        '200':
          description: Validation successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Account validated successfully"
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/profile:
    get:
      summary: Get authenticated user profile
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      email:
                        type: string
                        format: email
                      firstName:
                        type: string
                      lastName:
                        type: string
                      role:
                        type: string
                        example: "customer"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/admin/profile:
    get:
      summary: Get authenticated admin profile
      operationId: getAdminProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Admin profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Admin'
        '401':
          description: Unauthorized or not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/request-password-reset:
    post:
      summary: Request a password reset
      operationId: requestPasswordReset
      description: Sends a password reset email to the user's email address
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          description: Password reset email sent (success response given even if email doesn't exist for security)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "If the email exists in our system, a password reset link has been sent"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/reset-password:
    post:
      summary: Reset password using token
      operationId: resetPassword
      description: Resets a user's password using the token sent via email
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordReset'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Password has been reset successfully"
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'